[6] refactor extract validation module, settings module was getting too big
